extract reporters to their own files
