+++ packages/test/src/cli.ts
@@ -4,7 +4,7 @@ import type * as http from 'node:http'
 import * as path from 'node:path'
 import { tsImport } from 'tsx/esm/api'
 import { runServerTests } from './lib/runner.ts'
-import { createReporter } from './lib/reporter.ts'
+import { createReporter } from './lib/reporters/index.ts'
 import { createWatcher } from './lib/watcher.ts'
 import { loadPlaywrightConfig, resolveProjects } from './lib/playwright.ts'
 import { loadConfig, type ResolvedRemixTestConfig } from './lib/config.ts'
