+++ packages/test/src/lib/reporters/spec.ts
@@ -1,13 +1,6 @@
-import { colors, normalizeLine, type Counts } from './utils.ts'
-import type { TestResult, TestResults } from './executor.ts'
-
-export interface Reporter {
-  onResult(results: TestResults, env?: string): void
-  onSummary(counts: Counts, durationMs: number): void
-  onSectionStart(label: string): void
-}
-
-// ── Spec ─────────────────────────────────────────────────────────────────────
+import { colors, normalizeLine, type Counts } from '../utils.ts'
+import type { TestResult, TestResults } from '../executor.ts'
+import type { Reporter } from './index.ts'
 
 export class SpecReporter implements Reporter {
   #failures: { suiteName: string; name: string; error: TestResult['error'] }[] = []
@@ -176,132 +169,5 @@ export class SpecReporter implements Reporter {
     if (todo > 0) console.log(`${info} todo ${todo}`)
     console.log(`${info} duration_ms ${durationMs.toFixed(5)}`)
     console.log()
-  }
-}
-
-// ── TAP ──────────────────────────────────────────────────────────────────────
-
-export class TapReporter implements Reporter {
-  #counter = 0
-  #total = 0
-
-  onSectionStart(_label: string) {}
-
-  onResult(results: TestResults, env?: string) {
-    if (this.#counter === 0) {
-      console.log('TAP version 14')
-    }
-
-    let envComment = env ? ` # ${env}` : ''
-
-    for (let test of results.tests) {
-      this.#counter++
-      this.#total++
-      let fullName = test.name
-        ? `${test.suiteName} > ${test.name}${envComment}`
-        : `${test.suiteName}${envComment}`
-
-      if (test.status === 'passed') {
-        console.log(`ok ${this.#counter} - ${fullName}`)
-      } else if (test.status === 'skipped') {
-        console.log(`ok ${this.#counter} - ${fullName} # SKIP`)
-      } else if (test.status === 'todo') {
-        console.log(`ok ${this.#counter} - ${fullName} # TODO`)
-      } else {
-        console.log(`not ok ${this.#counter} - ${fullName}`)
-        console.log('  ---')
-        console.log(`  message: ${test.error?.message ?? 'unknown error'}`)
-        if (test.error?.stack) {
-          let frames = test.error.stack
-            .split('\n')
-            .slice(1, 4)
-            .map((l) => normalizeLine(l).trim())
-            .join('\n    ')
-          console.log(`  stack: |\n    ${frames}`)
-        }
-        console.log('  ...')
-      }
-    }
-  }
-
-  onSummary(counts: Counts, durationMs: number) {
-    let { passed, failed, skipped, todo } = counts
-    console.log(`1..${this.#total}`)
-    console.log(`# tests ${passed + failed + skipped + todo}`)
-    console.log(`# pass ${passed}`)
-    console.log(`# fail ${failed}`)
-    if (skipped > 0) console.log(`# skipped ${skipped}`)
-    if (todo > 0) console.log(`# todo ${todo}`)
-    console.log(`# duration_ms ${durationMs.toFixed(5)}`)
-  }
-}
-
-// ── Dot ──────────────────────────────────────────────────────────────────────
-
-export class DotReporter implements Reporter {
-  #failures: { name: string; error: TestResult['error'] }[] = []
-  #dotCount = 0
-
-  onSectionStart(_label: string) {}
-
-  onResult(results: TestResults, _env?: string) {
-    for (let test of results.tests) {
-      if (test.status === 'passed') {
-        process.stdout.write(colors.green('.'))
-      } else if (test.status === 'skipped') {
-        process.stdout.write(colors.dim('S'))
-      } else if (test.status === 'todo') {
-        process.stdout.write(colors.dim('T'))
-      } else {
-        process.stdout.write(colors.red('F'))
-        this.#failures.push({ name: `${test.suiteName} > ${test.name}`, error: test.error })
-      }
-      this.#dotCount++
-    }
-  }
-
-  onSummary(counts: Counts, durationMs: number) {
-    if (this.#dotCount > 0) console.log()
-
-    for (let i = 0; i < this.#failures.length; i++) {
-      let { name, error } = this.#failures[i]
-      console.log(`\n  ${colors.red(`${i + 1})`)} ${name}`)
-      if (error) {
-        console.log(`     ${colors.red(error.message)}`)
-        if (error.stack) {
-          let frames = error.stack
-            .split('\n')
-            .slice(1, 4)
-            .map((l) => `     ${normalizeLine(l).trim()}`)
-            .join('\n')
-          console.log(frames)
-        }
-      }
-    }
-
-    let { passed, failed, skipped, todo } = counts
-    let info = colors.cyan('ℹ')
-    console.log()
-    console.log(`${info} tests ${passed + failed + skipped + todo}`)
-    console.log(`${info} pass ${passed}`)
-    console.log(`${info} fail ${failed}`)
-    if (skipped > 0) console.log(`${info} skipped ${skipped}`)
-    if (todo > 0) console.log(`${info} todo ${todo}`)
-    console.log(`${info} duration_ms ${durationMs.toFixed(5)}`)
-    console.log()
-  }
-}
-
-// ── Factory ───────────────────────────────────────────────────────────────────
-
-export function createReporter(type: string): Reporter {
-  switch (type) {
-    case 'tap':
-      return new TapReporter()
-    case 'dot':
-      return new DotReporter()
-    case 'spec':
-    default:
-      return new SpecReporter()
   }
 }
