+++ packages/test/src/lib/runner.ts
@@ -3,7 +3,7 @@ import { pathToFileURL } from 'node:url'
 import { Worker } from 'node:worker_threads'
 import type { TestResults } from './executor.ts'
 import { type PlaywrightUseOpts } from './playwright.ts'
-import type { Reporter } from './reporter.ts'
+import type { Reporter } from './reporters/index.ts'
 import type { Counts } from './utils.ts'
 
 const ext = path.extname(import.meta.url)
